Create #5
In a puzzle contest, the chairman of your English club posts a problem to compare a given pair 
of words and eliminate all common characters in them. To speed up the process of judging, the 
computer club head was requested to prepare computer logic. Please code a solution to the 
above problem applying your own skillset.